--- src/main/java/io/swagger/api/ApiException.java
@@ -1,10 +1,0 @@
-package io.swagger.api;
-
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-11-21T13:18:37.550Z[GMT]")
-public class ApiException extends Exception {
-    private int code;
-    public ApiException (int code, String msg) {
-        super(msg);
-        this.code = code;
-    }
-}
