+++ src/main/java/io/swagger/exception/BadInputException.java
@@ -1,0 +1,2 @@
+package io.swagger.exception;public class BadInputException {
+}
